+++ docs/assets/UluTabs.stories-BEaKq6bN.js
@@ -1,0 +1,131 @@
+import{d as u,e as s,f as r,g as o,h as t}from"./iframe-DZKOIr3v.js";import"./preload-helper-BJwshlQW.js";const U={component:t,subcomponents:{UluTabList:o,UluTab:r,UluTabPanels:s,UluTabPanel:u},tags:["autodocs"]},l=i=>({components:{UluTabGroup:t,UluTabList:o,UluTab:r,UluTabPanels:s,UluTabPanel:u},setup(){return{args:i}},template:`
+    <UluTabGroup v-bind="args">
+      <UluTabList>
+        <UluTab>Profile</UluTab>
+        <UluTab>Settings</UluTab>
+        <UluTab>Notifications</UluTab>
+      </UluTabList>
+      <UluTabPanels>
+        <UluTabPanel>
+          <h3>User Profile</h3>
+          <p>This is where your profile information will be displayed.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Account Settings</h3>
+          <p>Manage your account preferences here.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Your Notifications</h3>
+          <p>You have 3 new notifications.</p>
+        </UluTabPanel>
+      </UluTabPanels>
+    </UluTabGroup>
+  `}),a=l.bind({});a.args={};const n=l.bind({});n.args={vertical:!0};const e=l.bind({});e.args={defaultIndex:2};a.parameters={...a.parameters,docs:{...a.parameters?.docs,source:{originalSource:`args => ({
+  components: {
+    UluTabGroup,
+    UluTabList,
+    UluTab,
+    UluTabPanels,
+    UluTabPanel
+  },
+  setup() {
+    return {
+      args
+    };
+  },
+  template: \`
+    <UluTabGroup v-bind="args">
+      <UluTabList>
+        <UluTab>Profile</UluTab>
+        <UluTab>Settings</UluTab>
+        <UluTab>Notifications</UluTab>
+      </UluTabList>
+      <UluTabPanels>
+        <UluTabPanel>
+          <h3>User Profile</h3>
+          <p>This is where your profile information will be displayed.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Account Settings</h3>
+          <p>Manage your account preferences here.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Your Notifications</h3>
+          <p>You have 3 new notifications.</p>
+        </UluTabPanel>
+      </UluTabPanels>
+    </UluTabGroup>
+  \`
+})`,...a.parameters?.docs?.source}}};n.parameters={...n.parameters,docs:{...n.parameters?.docs,source:{originalSource:`args => ({
+  components: {
+    UluTabGroup,
+    UluTabList,
+    UluTab,
+    UluTabPanels,
+    UluTabPanel
+  },
+  setup() {
+    return {
+      args
+    };
+  },
+  template: \`
+    <UluTabGroup v-bind="args">
+      <UluTabList>
+        <UluTab>Profile</UluTab>
+        <UluTab>Settings</UluTab>
+        <UluTab>Notifications</UluTab>
+      </UluTabList>
+      <UluTabPanels>
+        <UluTabPanel>
+          <h3>User Profile</h3>
+          <p>This is where your profile information will be displayed.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Account Settings</h3>
+          <p>Manage your account preferences here.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Your Notifications</h3>
+          <p>You have 3 new notifications.</p>
+        </UluTabPanel>
+      </UluTabPanels>
+    </UluTabGroup>
+  \`
+})`,...n.parameters?.docs?.source}}};e.parameters={...e.parameters,docs:{...e.parameters?.docs,source:{originalSource:`args => ({
+  components: {
+    UluTabGroup,
+    UluTabList,
+    UluTab,
+    UluTabPanels,
+    UluTabPanel
+  },
+  setup() {
+    return {
+      args
+    };
+  },
+  template: \`
+    <UluTabGroup v-bind="args">
+      <UluTabList>
+        <UluTab>Profile</UluTab>
+        <UluTab>Settings</UluTab>
+        <UluTab>Notifications</UluTab>
+      </UluTabList>
+      <UluTabPanels>
+        <UluTabPanel>
+          <h3>User Profile</h3>
+          <p>This is where your profile information will be displayed.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Account Settings</h3>
+          <p>Manage your account preferences here.</p>
+        </UluTabPanel>
+        <UluTabPanel>
+          <h3>Your Notifications</h3>
+          <p>You have 3 new notifications.</p>
+        </UluTabPanel>
+      </UluTabPanels>
+    </UluTabGroup>
+  \`
+})`,...e.parameters?.docs?.source}}};const p=["Default","Vertical","DefaultIndex"];export{a as Default,e as DefaultIndex,n as Vertical,p as __namedExportsOrder,U as default};