+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 0.1.1-beta.17
+
+- Create `lib/components/systems/facets/UluFacetsFilterAccordions.vue` for accordion implementation of filter lists, `UluFacetsFilterLists` is the unstyled collapsible version
+- `UluFacetsFilterLists` Remove old props for `UluCollapsible`
+
 ## 0.1.1-beta.16
 
 - `lib/components/systems/facets/UluFacetsActiveFilters`